refactor(TransferManagerWorker): Split query in a file
