+++ SwissTransferFeatures/UploadProgressView/TransferManagerWorker+Query.swift
@@ -1,0 +1,64 @@
+/*
+ Infomaniak SwissTransfer - iOS App
+ Copyright (C) 2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+import InfomaniakCore
+import InfomaniakDI
+import STCore
+import STNetwork
+import SwissTransferCore
+
+extension TransferManagerWorker {
+    func uploadChunk(
+        chunk: Data,
+        index: Int,
+        isLastChunk: Bool,
+        remoteUploadFileUUID: String,
+        uploadUUID: String
+    ) async throws {
+        @InjectService var injection: SwissTransferInjection
+        guard let rawChunkURL = try injection.sharedApiUrlCreator.uploadChunkUrl(
+            uploadUUID: uploadUUID,
+            fileUUID: remoteUploadFileUUID,
+            chunkIndex: Int32(index),
+            isLastChunk: isLastChunk,
+            isRetry: false
+        ) else {
+            throw TransferSessionManager.ErrorDomain.invalidUploadChunkURL
+        }
+
+        guard let chunkURL = URL(string: rawChunkURL) else {
+            throw TransferSessionManager.ErrorDomain.invalidURL(rawURL: rawChunkURL)
+        }
+
+        var uploadRequest = URLRequest(url: chunkURL)
+        uploadRequest.httpMethod = Method.POST.rawValue
+
+        let taskDelegate = UploadTaskDelegate(totalBytesExpectedToSend: chunk.count)
+        overallProgress.addChild(taskDelegate.taskProgress, withPendingUnitCount: Int64(chunk.count))
+
+        let (_, response) = try await uploadURLSession.upload(for: uploadRequest, from: chunk, delegate: taskDelegate)
+        guard let httpResponse = response as? HTTPURLResponse else {
+            throw TransferSessionManager.ErrorDomain.invalidResponse
+        }
+
+        if httpResponse.statusCode >= 400 {
+            throw TransferSessionManager.ErrorDomain.invalidChunkResponse
+        }
+    }
+}
